requirements/base.txt
Django==2.0.5
djangorestframework==3.7.3
coreapi==2.3.3
psycopg2-binary>=2.7,<2.8
boto3==1.7.19
django-boto==0.3.12
django-filter==1.0.4
pytz==2018.3
